feat: implement ShoppingItemRepository with firestore

+++ app/src/main/java/com/viniciusrodriguesaro/carry/shoppingitem/dto/ShoppingItem.kt
@@ -7,7 +7,7 @@ data class ShoppingItem(
     var isCompleted: Boolean,
     var name: String,
     var description: String,
-    var price: Double?,
+    var price: Int?,
     var amount: Int?,
     var unitOfMeasurement: MeasurementType?
 )
